+++ vms/platformvm/validators/manager_test.go
@@ -5,9 +5,6 @@ package validators
 
 import (
 	"context"
-	"encoding/hex"
-	"errors"
-	"fmt"
 	"testing"
 	"time"
 
@@ -18,66 +15,28 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ava-labs/avalanchego/chains"
-	"github.com/ava-labs/avalanchego/codec"
-	"github.com/ava-labs/avalanchego/codec/linearcodec"
 	"github.com/ava-labs/avalanchego/database"
-	dbmanager "github.com/ava-labs/avalanchego/database/manager"
-	"github.com/ava-labs/avalanchego/database/versiondb"
 	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/snow"
 	"github.com/ava-labs/avalanchego/snow/choices"
 	"github.com/ava-labs/avalanchego/snow/uptime"
 	"github.com/ava-labs/avalanchego/snow/validators"
-	"github.com/ava-labs/avalanchego/utils"
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/utils/crypto/bls"
-	"github.com/ava-labs/avalanchego/utils/crypto/secp256k1"
-	"github.com/ava-labs/avalanchego/utils/formatting"
-	"github.com/ava-labs/avalanchego/utils/formatting/address"
-	"github.com/ava-labs/avalanchego/utils/json"
-	"github.com/ava-labs/avalanchego/utils/logging"
-	"github.com/ava-labs/avalanchego/utils/set"
 	"github.com/ava-labs/avalanchego/utils/timer/mockable"
 	"github.com/ava-labs/avalanchego/utils/units"
-	"github.com/ava-labs/avalanchego/version"
-	"github.com/ava-labs/avalanchego/vms/components/avax"
-	"github.com/ava-labs/avalanchego/vms/platformvm/api"
 	"github.com/ava-labs/avalanchego/vms/platformvm/blocks"
 	"github.com/ava-labs/avalanchego/vms/platformvm/config"
 	"github.com/ava-labs/avalanchego/vms/platformvm/metrics"
 	"github.com/ava-labs/avalanchego/vms/platformvm/reward"
-	"github.com/ava-labs/avalanchego/vms/platformvm/signer"
 	"github.com/ava-labs/avalanchego/vms/platformvm/state"
-	"github.com/ava-labs/avalanchego/vms/platformvm/status"
-	"github.com/ava-labs/avalanchego/vms/platformvm/txs"
-	"github.com/ava-labs/avalanchego/vms/platformvm/txs/builder"
-	"github.com/ava-labs/avalanchego/vms/platformvm/utxo"
-	"github.com/ava-labs/avalanchego/vms/secp256k1fx"
 )
 
-const (
-	defaultTxFee              = uint64(100)
-	defaultMinStakingDuration = 24 * time.Hour
-	defaultMaxStakingDuration = 365 * 24 * time.Hour
-	defaultMinValidatorStake  = 5 * units.MilliAvax
-	defaultBalance            = 100 * defaultMinValidatorStake
-)
-
-// AVAX asset ID in tests
-var (
-	defaultRewardConfig = reward.Config{
-		MaxConsumptionRate: .12 * reward.PercentDenominator,
-		MinConsumptionRate: .10 * reward.PercentDenominator,
-		MintingPeriod:      365 * 24 * time.Hour,
-		SupplyCap:          720 * units.MegaAvax,
-	}
-
-	defaultGenesisTime = time.Date(1997, 1, 1, 0, 0, 0, 0, time.UTC)
-	preFundedKeys      = secp256k1.TestKeys()
-	avaxAssetID        = ids.ID{'y', 'e', 'e', 't'}
-	xChainID           = ids.Empty.Prefix(0)
-	cChainID           = ids.Empty.Prefix(1)
-)
+var defaultRewardConfig = reward.Config{
+	MaxConsumptionRate: .12 * reward.PercentDenominator,
+	MinConsumptionRate: .10 * reward.PercentDenominator,
+	MintingPeriod:      365 * 24 * time.Hour,
+	SupplyCap:          720 * units.MegaAvax,
+}
 
 func TestVM_GetValidatorSet(t *testing.T) {
 	// Populate the validator set to use below
@@ -471,453 +430,6 @@ func TestVM_GetValidatorSet(t *testing.T) {
 			r.Equal(tt.expectedVdrSet, gotVdrSet)
 		})
 	}
-}
-
-func Test_RegressionBLSKeyDiff(t *testing.T) {
-	// setup
-	require := require.New(t)
-	cfg, ctx, clk, pState, txBuilder, err := buildSetup()
-	require.NoError(err)
-
-	currentHeight := uint64(1)
-	currentTime := clk.Time().Add(time.Second)
-	clk.Set(currentTime)
-
-	// add a subnet
-	subnetTx, err := txBuilder.NewCreateSubnetTx(
-		1, // threshold
-		[]ids.ShortID{ // control keys
-			preFundedKeys[0].PublicKey().Address(),
-		},
-		[]*secp256k1.PrivateKey{preFundedKeys[0]},
-		preFundedKeys[0].PublicKey().Address(),
-	)
-	require.NoError(err)
-	pState.AddSubnet(subnetTx)
-	pState.AddTx(subnetTx, status.Committed)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(clk.Time())
-	require.NoError(pState.Commit())
-
-	subnetID := subnetTx.ID()
-	cfg.TrackedSubnets.Add(subnetID)
-	subnetValidators := validators.NewSet()
-	require.NoError(pState.ValidatorSet(subnetID, subnetValidators))
-	require.True(cfg.Validators.Add(subnetID, subnetValidators))
-
-	// A subnet validator should stake twice before its primary network counterpart stops staking
-	var (
-		primaryStartTime   = currentTime.Add(time.Second)
-		primaryStartHeight = currentHeight + 1
-
-		subnetStartTime1   = primaryStartTime.Add(2 * time.Second)
-		subnetStartHeight1 = primaryStartHeight + 1
-
-		subnetEndTime1   = subnetStartTime1.Add(defaultMinStakingDuration)
-		subnetEndHeight1 = subnetStartHeight1 + 1
-
-		subnetStartTime2   = subnetEndTime1.Add(2 * time.Second)
-		subnetStartHeight2 = subnetEndHeight1 + 1
-
-		subnetEndTime2   = subnetStartTime2.Add(defaultMinStakingDuration)
-		subnetEndHeight2 = subnetStartHeight2 + 1
-
-		primaryEndTime = subnetEndTime2.Add(time.Second)
-	)
-
-	// move time/height ahead
-	currentHeight = primaryStartHeight
-	currentTime = primaryStartTime
-	clk.Set(currentTime)
-
-	// insert primary network validator
-	nodeID := ids.GenerateTestNodeID()
-	addr := preFundedKeys[0].PublicKey().Address()
-	skBytes, err := hex.DecodeString("6668fecd4595b81e4d568398c820bbf3f073cb222902279fa55ebb84764ed2e3")
-	require.NoError(err)
-	sk, err := bls.SecretKeyFromBytes(skBytes)
-	require.NoError(err)
-
-	uPrimaryTx := &txs.AddPermissionlessValidatorTx{
-		BaseTx: txs.BaseTx{BaseTx: avax.BaseTx{
-			NetworkID:    ctx.NetworkID,
-			BlockchainID: ctx.ChainID,
-			Ins:          []*avax.TransferableInput{},
-			Outs:         []*avax.TransferableOutput{},
-		}},
-		Validator: txs.Validator{
-			NodeID: nodeID,
-			Start:  uint64(primaryStartTime.Unix()),
-			End:    uint64(primaryEndTime.Unix()),
-			Wght:   2000000000000,
-		},
-		Subnet: constants.PrimaryNetworkID,
-		Signer: signer.NewProofOfPossession(sk),
-		StakeOuts: []*avax.TransferableOutput{
-			{
-				Asset: avax.Asset{
-					ID: avaxAssetID,
-				},
-				Out: &secp256k1fx.TransferOutput{
-					Amt: 2 * units.KiloAvax,
-					OutputOwners: secp256k1fx.OutputOwners{
-						Locktime:  0,
-						Threshold: 1,
-						Addrs: []ids.ShortID{
-							addr,
-						},
-					},
-				},
-			},
-		},
-		ValidatorRewardsOwner: &secp256k1fx.OutputOwners{
-			Locktime:  0,
-			Threshold: 1,
-			Addrs: []ids.ShortID{
-				addr,
-			},
-		},
-		DelegatorRewardsOwner: &secp256k1fx.OutputOwners{
-			Locktime:  0,
-			Threshold: 1,
-			Addrs: []ids.ShortID{
-				addr,
-			},
-		},
-		DelegationShares: reward.PercentDenominator,
-	}
-	primaryTx, err := txs.NewSigned(uPrimaryTx, txs.Codec, nil)
-	require.NoError(err)
-	require.NoError(primaryTx.SyntacticVerify(ctx))
-
-	require.NoError(err)
-	dummyBlock, err := blocks.NewBanffStandardBlock(
-		currentTime,
-		pState.GetLastAccepted(),
-		currentHeight,
-		[]*txs.Tx{primaryTx},
-	)
-	require.NoError(err)
-
-	primaryValidator, err := state.NewCurrentStaker(
-		primaryTx.ID(),
-		uPrimaryTx,
-		0,
-	)
-	require.NoError(err)
-
-	pState.PutCurrentValidator(primaryValidator)
-	pState.SetLastAccepted(dummyBlock.ID())
-	pState.AddStatelessBlock(dummyBlock, choices.Accepted)
-	pState.AddTx(primaryTx, status.Committed)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(currentTime)
-	require.NoError(pState.Commit())
-
-	// move time/height ahead
-	currentHeight = subnetStartHeight1
-	currentTime = subnetStartTime1
-	clk.Set(currentTime)
-
-	// insert first subnet validator
-	subnetFirstTx, err := txBuilder.NewAddSubnetValidatorTx(
-		1,                               // Weight
-		uint64(subnetStartTime1.Unix()), // Start time
-		uint64(subnetEndTime1.Unix()),   // end time
-		nodeID,                          // Node ID
-		subnetID,
-		[]*secp256k1.PrivateKey{preFundedKeys[0]},
-		addr,
-	)
-	require.NoError(err)
-	dummyBlock, err = blocks.NewBanffStandardBlock(
-		currentTime,
-		pState.GetLastAccepted(),
-		currentHeight,
-		[]*txs.Tx{subnetFirstTx},
-	)
-	require.NoError(err)
-
-	firstSubnetValidator, err := state.NewCurrentStaker(
-		subnetFirstTx.ID(),
-		subnetFirstTx.Unsigned.(*txs.AddSubnetValidatorTx),
-		0,
-	)
-	require.NoError(err)
-
-	pState.PutCurrentValidator(firstSubnetValidator)
-	pState.SetLastAccepted(dummyBlock.ID())
-	pState.AddStatelessBlock(dummyBlock, choices.Accepted)
-	pState.AddTx(subnetFirstTx, status.Committed)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(currentTime)
-	require.NoError(pState.Commit())
-
-	// move time/height ahead
-	currentHeight = subnetEndHeight1
-	currentTime = subnetEndTime1
-	clk.Set(currentTime)
-
-	// drop first subnet validator
-	dummyBlock, err = blocks.NewBanffStandardBlock(
-		currentTime,
-		pState.GetLastAccepted(),
-		currentHeight,
-		[]*txs.Tx{},
-	)
-	require.NoError(err)
-
-	pState.DeleteCurrentValidator(firstSubnetValidator)
-	pState.SetLastAccepted(dummyBlock.ID())
-	pState.AddStatelessBlock(dummyBlock, choices.Accepted)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(currentTime)
-	require.NoError(pState.Commit())
-
-	// move time/height ahead
-	currentHeight = subnetStartHeight2
-	currentTime = subnetStartTime2
-	clk.Set(currentTime)
-
-	// insert second subnet validator
-	subnetSecondTx, err := txBuilder.NewAddSubnetValidatorTx(
-		1,                               // Weight
-		uint64(subnetStartTime2.Unix()), // Start time
-		uint64(subnetEndTime2.Unix()),   // end time
-		nodeID,                          // Node ID
-		subnetID,
-		[]*secp256k1.PrivateKey{preFundedKeys[0]},
-		addr,
-	)
-	require.NoError(err)
-	dummyBlock, err = blocks.NewBanffStandardBlock(
-		currentTime,
-		pState.GetLastAccepted(),
-		currentHeight,
-		[]*txs.Tx{subnetSecondTx},
-	)
-	require.NoError(err)
-
-	secondSubnetValidator, err := state.NewCurrentStaker(
-		subnetSecondTx.ID(),
-		subnetSecondTx.Unsigned.(*txs.AddSubnetValidatorTx),
-		0,
-	)
-	require.NoError(err)
-
-	pState.PutCurrentValidator(secondSubnetValidator)
-	pState.SetLastAccepted(dummyBlock.ID())
-	pState.AddStatelessBlock(dummyBlock, choices.Accepted)
-	pState.AddTx(subnetSecondTx, status.Committed)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(currentTime)
-	require.NoError(pState.Commit())
-
-	// move time/height ahead
-	currentHeight = subnetEndHeight2
-	currentTime = subnetEndTime2
-	clk.Set(currentTime)
-
-	// drop second subnet validator
-	dummyBlock, err = blocks.NewBanffStandardBlock(
-		currentTime,
-		pState.GetLastAccepted(),
-		currentHeight,
-		[]*txs.Tx{},
-	)
-	require.NoError(err)
-
-	pState.DeleteCurrentValidator(secondSubnetValidator)
-	pState.SetLastAccepted(dummyBlock.ID())
-	pState.AddStatelessBlock(dummyBlock, choices.Accepted)
-	pState.SetHeight(currentHeight)
-	pState.SetTimestamp(currentTime)
-	require.NoError(pState.Commit())
-
-	// Finally the test
-	valMan := NewManager(*cfg, pState, metrics.Noop, clk)
-	for _, height := range []uint64{
-		primaryStartHeight,
-		subnetStartHeight1,
-		subnetEndHeight2,
-	} {
-		primaryVals, err := valMan.GetValidatorSet(context.Background(), height, constants.PrimaryNetworkID)
-		require.NoError(err)
-
-		primaryVal, found := primaryVals[nodeID]
-		require.True(found)
-		require.NotNil(primaryVal.PublicKey)
-	}
-
-	for _, height := range []uint64{
-		subnetStartHeight1,
-		subnetStartHeight2,
-	} {
-		subnetVals, err := valMan.GetValidatorSet(context.Background(), height, subnetID)
-		require.NoError(err)
-
-		subnetVal, found := subnetVals[nodeID]
-		require.True(found)
-		require.NotNil(subnetVal.PublicKey)
-	}
-}
-
-func buildSetup() (
-	*config.Config,
-	*snow.Context,
-	*mockable.Clock,
-	state.State,
-	builder.Builder,
-	error,
-) {
-	// build platformVM config
-	vdrs := validators.NewManager()
-	primaryVdrs := validators.NewSet()
-	_ = vdrs.Add(constants.PrimaryNetworkID, primaryVdrs)
-	trackedSubnets := set.NewSet[ids.ID](1) // will add subnet later on
-	cfg := &config.Config{
-		Chains:                 chains.TestManager,
-		UptimeLockedCalculator: uptime.NewLockedCalculator(),
-		Validators:             vdrs,
-		TrackedSubnets:         trackedSubnets,
-		TxFee:                  defaultTxFee,
-		CreateSubnetTxFee:      100 * defaultTxFee,
-		CreateBlockchainTxFee:  100 * defaultTxFee,
-		MinValidatorStake:      5 * units.MilliAvax,
-		MaxValidatorStake:      500 * units.MilliAvax,
-		MinDelegatorStake:      1 * units.MilliAvax,
-		MinStakeDuration:       defaultMinStakingDuration,
-		MaxStakeDuration:       defaultMaxStakingDuration,
-		RewardConfig:           defaultRewardConfig,
-	}
-
-	// build context
-	ctx := snow.DefaultContextTest()
-	ctx.NetworkID = 10
-	ctx.XChainID = xChainID
-	ctx.CChainID = cChainID
-	ctx.AVAXAssetID = avaxAssetID
-	ctx.ValidatorState = &validators.TestState{
-		GetSubnetIDF: func(_ context.Context, chainID ids.ID) (ids.ID, error) {
-			subnetID, ok := map[ids.ID]ids.ID{
-				constants.PlatformChainID: constants.PrimaryNetworkID,
-				xChainID:                  constants.PrimaryNetworkID,
-				cChainID:                  constants.PrimaryNetworkID,
-			}[chainID]
-			if !ok {
-				return ids.Empty, errors.New("missing")
-			}
-			return subnetID, nil
-		},
-	}
-	clk := &mockable.Clock{}
-	clk.Set(defaultGenesisTime)
-
-	baseDBManager := dbmanager.NewMemDB(version.Semantic1_0_0)
-	db := versiondb.New(baseDBManager.Current().Database)
-	rewardsCalc := reward.NewCalculator(cfg.RewardConfig)
-	genesisBytes := buildGenesisTest(ctx)
-	pState, err := state.New(
-		db,
-		genesisBytes,
-		prometheus.NewRegistry(),
-		cfg,
-		ctx,
-		metrics.Noop,
-		rewardsCalc,
-		&utils.Atomic[bool]{},
-	)
-	if err != nil {
-		return nil, nil, nil, nil, nil, fmt.Errorf("could not build pState, %w", err)
-	}
-
-	// build fx
-	fxVMInt := &fxVMInt{
-		registry: linearcodec.NewDefault(),
-		clk:      clk,
-		log:      ctx.Log,
-	}
-	fx := &secp256k1fx.Fx{}
-	if err := fx.Initialize(fxVMInt); err != nil {
-		return nil, nil, nil, nil, nil, fmt.Errorf("could not build feature extension, %w", err)
-	}
-	if err := fx.Bootstrapped(); err != nil {
-		return nil, nil, nil, nil, nil, fmt.Errorf("could not bootstrap feature extension, %w", err)
-	}
-
-	atomicUTXOs := avax.NewAtomicUTXOManager(ctx.SharedMemory, txs.Codec)
-	utxoHandler := utxo.NewHandler(ctx, clk, fx)
-	txBuilder := builder.New(
-		ctx,
-		cfg,
-		clk,
-		fx,
-		pState,
-		atomicUTXOs,
-		utxoHandler,
-	)
-	return cfg, ctx, clk, pState, txBuilder, nil
-}
-
-type fxVMInt struct {
-	registry codec.Registry
-	clk      *mockable.Clock
-	log      logging.Logger
-}
-
-func (fvi *fxVMInt) CodecRegistry() codec.Registry {
-	return fvi.registry
-}
-
-func (fvi *fxVMInt) Clock() *mockable.Clock {
-	return fvi.clk
-}
-
-func (fvi *fxVMInt) Logger() logging.Logger {
-	return fvi.log
-}
-
-func buildGenesisTest(ctx *snow.Context) []byte {
-	genesisUTXOs := make([]api.UTXO, len(preFundedKeys))
-	for i, key := range preFundedKeys {
-		id := key.PublicKey().Address()
-		addr, err := address.FormatBech32(constants.UnitTestHRP, id.Bytes())
-		if err != nil {
-			panic(err)
-		}
-		genesisUTXOs[i] = api.UTXO{
-			Amount:  json.Uint64(defaultBalance),
-			Address: addr,
-		}
-	}
-
-	// No validators in this genesis. We want to control in test which stakers are added and when
-	genesisValidators := make([]api.PermissionlessValidator, 0)
-	defaultGenesisTime := time.Date(1997, 1, 1, 0, 0, 0, 0, time.UTC)
-
-	buildGenesisArgs := api.BuildGenesisArgs{
-		NetworkID:     json.Uint32(constants.UnitTestID),
-		AvaxAssetID:   ctx.AVAXAssetID,
-		UTXOs:         genesisUTXOs,
-		Validators:    genesisValidators,
-		Chains:        nil,
-		Time:          json.Uint64(defaultGenesisTime.Unix()),
-		InitialSupply: json.Uint64(360 * units.MegaAvax),
-		Encoding:      formatting.Hex,
-	}
-
-	buildGenesisResponse := api.BuildGenesisReply{}
-	platformvmSS := api.StaticService{}
-	if err := platformvmSS.BuildGenesis(nil, &buildGenesisArgs, &buildGenesisResponse); err != nil {
-		panic(fmt.Errorf("problem while building platform chain's genesis state: %w", err))
-	}
-
-	genesisBytes, err := formatting.Decode(buildGenesisResponse.Encoding, buildGenesisResponse.Bytes)
-	if err != nil {
-		panic(err)
-	}
-
-	return genesisBytes
 }
 
 func copyPrimaryValidator(vdr *validators.Validator) *validators.Validator {
